requirements.txt
langchain==0.0.267
langchain_openai==0.0.2
openai==1.3.0
anthropic==0.5.0
requests==2.31.0
streamlit==1.27.0
streamlit-chat==0.1.1
tiktoken==0.5.1
qdrant-client==1.6.0
PyPDF2==3.0.1
python-docx==0.8.11
pytesseract==0.3.10
Pillow==10.0.1
